chore(reg): Move the composition feature flag to a dedicated class (accessible from the Composition namespace!)
